CRUDV Lesson - First Draft

+++ Core API/Team7/Controllers/LessonController.cs
@@ -106,17 +106,14 @@ namespace Team7.Controllers
             bool assFlag = false;
 
             //check if attatched to employee:
-            //if (lesson.Employee != null)
-            //    assFlag = true;
-            //if (lesson.Schedule.Count != 0)
-            //    assFlag = true;
-            //if (lesson.LessonPlan.Count != 0)
-            //    assFlag = true;
+            if (lesson.Schedule.Count != 0)
+                assFlag = true;
 
-            //if (assFlag)
-            //    return StatusCode(StatusCodes.Status409Conflict, new { message = "Lesson has assosciations to another table.", lesson });
+            if (assFlag)
+                return StatusCode(StatusCodes.Status409Conflict, new { message = "Lesson has assosciations to another table.", lesson });
 
             //perform delete:
+            await _lessonPlanRepo.RemoveRangeLessonIdAsync(id);
             _lessonRepo.Delete(lesson);
             await _lessonRepo.SaveChangesAsync();
